# executor: Improve the performance of `aggFuncAvg` by using sliding window