--- a/executor/aggfuncs/func_avg.go
+++ b/executor/aggfuncs/func_avg.go
@@ -102,6 +102,48 @@ func (e *avgOriginal4Decimal) UpdatePartialResult(sctx sessionctx.Context, rowsI
 	return nil
 }
 
+func (e *avgOriginal4Decimal) Slide(sctx sessionctx.Context, rows []chunk.Row, lastStart, lastEnd uint64, shiftStart, shiftEnd uint64, pr PartialResult) error {
+	p := (*partialResult4AvgDecimal)(pr)
+	for i := uint64(0); i < shiftEnd; i++ {
+		input, isNull, err := e.args[0].EvalDecimal(sctx, rows[lastEnd+i])
+		if err != nil {
+			return err
+		}
+		if isNull {
+			continue
+		}
+		if p.count == 0 {
+			p.sum = *input
+			p.count = 1
+			continue
+		}
+		newSum := new(types.MyDecimal)
+		err = types.DecimalAdd(&p.sum, input, newSum)
+		if err != nil {
+			return err
+		}
+		p.sum = *newSum
+		p.count++
+	}
+	for i := uint64(0); i < shiftStart; i++ {
+		input, isNull, err := e.args[0].EvalDecimal(sctx, rows[lastStart+i])
+		if err != nil {
+			return err
+		}
+		if isNull {
+			continue
+		}
+		newSum := new(types.MyDecimal)
+		err = types.DecimalSub(&p.sum, input, newSum)
+		if err != nil {
+			return err
+		}
+		p.sum = *newSum
+		p.count--
+	}
+	return nil
+}
+
 type avgPartial4Decimal struct {
 	baseAvgDecimal
 }
@@ -132,6 +174,62 @@ func (e *avgPartial4Decimal) UpdatePartialResult(sctx sessionctx.Context, rowsIn
 		}
 		p.sum = *newSum
 		p.count += inputCount
+	}
+	return nil
+}
+
+func (e *avgPartial4Decimal) Slide(sctx sessionctx.Context, rows []chunk.Row, lastStart, lastEnd uint64, shiftStart, shiftEnd uint64, pr PartialResult) error {
+	p := (*partialResult4AvgDecimal)(pr)
+	for i := uint64(0); i < shiftEnd; i++ {
+		input, isNull, err := e.args[1].EvalDecimal(sctx, rows[lastEnd+i])
+		if err != nil {
+			return err
+		}
+		if isNull {
+			continue
+		}
+		inputCount, isNull, err := e.args[0].EvalInt(sctx, rows[lastEnd+i])
+		if err != nil {
+			return err
+		}
+		if isNull {
+			continue
+		}
+		if p.count == 0 {
+			p.sum = *input
+			p.count = inputCount
+			continue
+		}
+		newSum := new(types.MyDecimal)
+		err = types.DecimalAdd(&p.sum, input, newSum)
+		if err != nil {
+			return err
+		}
+		p.sum = *newSum
+		p.count += inputCount
+	}
+	for i := uint64(0); i < shiftStart; i++ {
+		input, isNull, err := e.args[1].EvalDecimal(sctx, rows[lastStart+i])
+		if err != nil {
+			return err
+		}
+		if isNull {
+			continue
+		}
+		inputCount, isNull, err := e.args[0].EvalInt(sctx, rows[lastStart+i])
+		if err != nil {
+			return err
+		}
+		if isNull {
+			continue
+		}
+		newSum := new(types.MyDecimal)
+		err = types.DecimalSub(&p.sum, input, newSum)
+		if err != nil {
+			return err
+		}
+		p.sum = *newSum
+		p.count -= inputCount
 	}
 	return nil
 }
@@ -264,11 +362,11 @@ func (e *baseAvgFloat64) AppendFinalResult2Chunk(sctx sessionctx.Context, pr Par
 	return nil
 }
 
-type avgOriginal4Float64 struct {
+type avgOriginal4Float64HighPrecision struct {
 	baseAvgFloat64
 }
 
-func (e *avgOriginal4Float64) UpdatePartialResult(sctx sessionctx.Context, rowsInGroup []chunk.Row, pr PartialResult) error {
+func (e *avgOriginal4Float64HighPrecision) UpdatePartialResult(sctx sessionctx.Context, rowsInGroup []chunk.Row, pr PartialResult) error {
 	p := (*partialResult4AvgFloat64)(pr)
 	for _, row := range rowsInGroup {
 		input, isNull, err := e.args[0].EvalReal(sctx, row)
@@ -285,11 +383,42 @@ func (e *avgOriginal4Float64) UpdatePartialResult(sctx sessionctx.Context, rowsI
 	return nil
 }
 
-type avgPartial4Float64 struct {
+type avgOriginal4Float64 struct {
+	avgOriginal4Float64HighPrecision
+}
+
+func (e *avgOriginal4Float64) Slide(sctx sessionctx.Context, rows []chunk.Row, lastStart, lastEnd uint64, shiftStart, shiftEnd uint64, pr PartialResult) error {
+	p := (*partialResult4AvgFloat64)(pr)
+	for i := uint64(0); i < shiftEnd; i++ {
+		input, isNull, err := e.args[0].EvalReal(sctx, rows[lastEnd+i])
+		if err != nil {
+			return err
+		}
+		if isNull {
+			continue
+		}
+		p.sum += input
+		p.count++
+	}
+	for i := uint64(0); i < shiftStart; i++ {
+		input, isNull, err := e.args[0].EvalReal(sctx, rows[lastStart+i])
+		if err != nil {
+			return err
+		}
+		if isNull {
+			continue
+		}
+		p.sum -= input
+		p.count--
+	}
+	return nil
+}
+
+type avgPartial4Float64HighPrecision struct {
 	baseAvgFloat64
 }
 
-func (e *avgPartial4Float64) UpdatePartialResult(sctx sessionctx.Context, rowsInGroup []chunk.Row, pr PartialResult) error {
+func (e *avgPartial4Float64HighPrecision) UpdatePartialResult(sctx sessionctx.Context, rowsInGroup []chunk.Row, pr PartialResult) error {
 	p := (*partialResult4AvgFloat64)(pr)
 	for _, row := range rowsInGroup {
 		inputSum, isNull, err := e.args[1].EvalReal(sctx, row)
@@ -313,10 +442,55 @@ func (e *avgPartial4Float64) UpdatePartialResult(sctx sessionctx.Context, rowsIn
 	return nil
 }
 
-func (e *avgPartial4Float64) MergePartialResult(sctx sessionctx.Context, src PartialResult, dst PartialResult) error {
+func (e *avgPartial4Float64HighPrecision) MergePartialResult(sctx sessionctx.Context, src PartialResult, dst PartialResult) error {
 	p1, p2 := (*partialResult4AvgFloat64)(src), (*partialResult4AvgFloat64)(dst)
 	p2.sum += p1.sum
 	p2.count += p1.count
+	return nil
+}
+
+type avgPartial4Float64 struct {
+	avgPartial4Float64HighPrecision
+}
+
+func (e *avgPartial4Float64) Slide(sctx sessionctx.Context, rows []chunk.Row, lastStart, lastEnd uint64, shiftStart, shiftEnd uint64, pr PartialResult) error {
+	p := (*partialResult4AvgFloat64)(pr)
+	for i := uint64(0); i < shiftEnd; i++ {
+		input, isNull, err := e.args[1].EvalReal(sctx, rows[lastEnd+i])
+		if err != nil {
+			return err
+		}
+		if isNull {
+			continue
+		}
+		inputCount, isNull, err := e.args[0].EvalInt(sctx, rows[lastEnd+i])
+		if err != nil {
+			return err
+		}
+		if isNull {
+			continue
+		}
+		p.sum += input
+		p.count += inputCount
+	}
+	for i := uint64(0); i < shiftStart; i++ {
+		input, isNull, err := e.args[1].EvalReal(sctx, rows[lastStart+i])
+		if err != nil {
+			return err
+		}
+		if isNull {
+			continue
+		}
+		inputCount, isNull, err := e.args[0].EvalInt(sctx, rows[lastStart+i])
+		if err != nil {
+			return err
+		}
+		if isNull {
+			continue
+		}
+		p.sum -= input
+		p.count -= inputCount
+	}
 	return nil
 }