### What problem does this PR solve?

Problem Summary: Improve the performance of  the aggFuncAvg

### What is changed and how it works?
[new.txt](https://github.com/pingcap/tidb/files/4370123/new.txt)
[old.txt](https://github.com/pingcap/tidb/files/4370124/old.txt)

related #14887

### Related changes

- PR to update `pingcap/docs`/`pingcap/docs-cn`: 
- PR to update `pingcap/tidb-ansible`:

I need add a variable of the `windowing_use_high_precision`

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

- Unit test

Side effects

* benchmark
```
name                                                                                                                                      old time/op  new time/op  delta

rows
WindowFunctionsWithSlidingWindow/(func:avg,_aggColType:float,_numFunc:1,_ndv:100,_rows:100000,_sorted:true,_concurrency:1)-16             19.7ms ± 1%   6.1ms ± 6%  -68.95%  (p=0.000 n=10+10)
WindowFunctionsWithSlidingWindow/(func:avg,_aggColType:decimal(11,0),_numFunc:1,_ndv:100,_rows:100000,_sorted:true,_concurrency:1)-16     92.6ms ± 1%  30.8ms ± 0%  -66.72%  (p=0.000 n=10+10)

ranges
WindowFunctionsWithSlidingWindow/(func:avg,_aggColType:float,_numFunc:1,_ndv:100,_rows:100000,_sorted:true,_concurrency:1)#01-16           698ms ± 1%     7ms ± 6%  -99.04%  (p=0.000 n=10+10)
WindowFunctionsWithSlidingWindow/(func:avg,_aggColType:decimal(11,0),_numFunc:1,_ndv:100,_rows:100000,_sorted:true,_concurrency:1)#01-16   3.19s ± 3%   0.03s ± 0%  -99.18%  (p=0.000 n=10+9)
```